Refactor- [:feature:user-profile] Apply & Fix Detekt, Ktlint Rules
Jira Tasks - [MM-81](https://mifosforge.jira.com/browse/MM-81)
