+++ androidApp/src/main/java/org/mifos/mobile/navigation/MifosNavGraph.kt
@@ -50,8 +50,6 @@ import org.mifos.mobile.feature.qr.navigation.navigateToQrDisplayScreen
 import org.mifos.mobile.feature.qr.navigation.navigateToQrImportScreen
 import org.mifos.mobile.feature.qr.navigation.navigateToQrReaderScreen
 import org.mifos.mobile.feature.qr.navigation.qrNavGraph
-import org.mifos.mobile.feature.transaction.navigation.navigateToRecentTransaction
-import org.mifos.mobile.feature.transaction.navigation.recentTransactionNavGraph
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsApplicationScreen
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsDetailScreen
 import org.mifos.mobile.feature.savings.navigation.navigateToSavingsMakeTransfer
@@ -60,12 +58,14 @@ import org.mifos.mobile.feature.settings.navigation.navigateToSettings
 import org.mifos.mobile.feature.settings.navigation.settingsNavGraph
 import org.mifos.mobile.feature.third.party.transfer.third_party_transfer.navigation.navigateToThirdPartyTransfer
 import org.mifos.mobile.feature.third.party.transfer.third_party_transfer.navigation.thirdPartyTransferNavGraph
+import org.mifos.mobile.feature.transaction.navigation.navigateToRecentTransaction
+import org.mifos.mobile.feature.transaction.navigation.recentTransactionNavGraph
 import org.mifos.mobile.feature.transfer.process.navigation.navigateToTransferProcessScreen
 import org.mifos.mobile.feature.transfer.process.navigation.transferProcessNavGraph
 import org.mifos.mobile.feature.update_password.navigation.navigateToUpdatePassword
 import org.mifos.mobile.feature.update_password.navigation.updatePasswordNavGraph
-import org.mifos.mobile.feature.user_profile.navigation.navigateToUserProfile
-import org.mifos.mobile.feature.user_profile.navigation.userProfileNavGraph
+import org.mifos.mobile.feature.user.profile.navigation.navigateToUserProfile
+import org.mifos.mobile.feature.user.profile.navigation.userProfileNavGraph
 import org.mifos.mobile.ui.activities.HomeActivity
 import org.mifos.mobile.ui.activities.PassCodeActivity
 
@@ -105,7 +105,7 @@ fun RootNavGraph(
         )
 
         userProfileNavGraph(
-            navController = navController,
+            navigateBack = navController::popBackStack,
             navigateToChangePassword = navController::navigateToUpdatePassword
         )
