+++ feature/user-profile/build.gradle.kts
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2024 Mifos Initiative
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ *
+ * See https://github.com/openMF/mobile-mobile/blob/master/LICENSE.md
+ */
+plugins {
+    alias(libs.plugins.mifos.android.feature)
+    alias(libs.plugins.mifos.android.library.compose)
+    alias(libs.plugins.kotlin.parcelize)
+}
+
+
+android {
+    namespace = "org.mifos.mobile.feature.user.profile"
+}
+
+dependencies {
+    implementation(libs.squareup.okhttp)
+}
