+++ feature/user-profile/src/main/res/drawable/mifos_logo.png
