--- feature/user_profile/build.gradle.kts
@@ -1,29 +1,0 @@
-plugins {
-    alias(libs.plugins.mifos.android.feature)
-    alias(libs.plugins.mifos.android.library.compose)
-    alias(libs.plugins.kotlin.parcelize)
-}
-
-
-android {
-    namespace = "org.mifos.mobile.feature.third.party.user_profile"
-}
-
-dependencies {
-    implementation(projects.core.ui)
-    implementation(projects.core.common)
-    implementation(projects.core.model)
-    implementation(projects.core.data)
-    
-    implementation(libs.squareup.retrofit2) {
-        // exclude Retrofit’s OkHttp peer-dependency module and define your own module import
-        exclude(module = "okhttp")
-    }
-    implementation(libs.squareup.retrofit.adapter.rxjava)
-    implementation(libs.squareup.retrofit.converter.gson)
-    implementation(libs.squareup.okhttp)
-    implementation(libs.squareup.logging.interceptor)
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.androidx.test.ext.junit)
-    androidTestImplementation(libs.espresso.core)
-}
