--- feature/user_profile/src/main/res/drawable/ic_gender_24dp.xml
@@ -1,45 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m266.859,466.778h-21.719c-3.952,0 -7.155,-3.203 -7.155,-7.155v-407.247c0,-3.952 3.203,-7.155 7.155,-7.155h21.719c3.952,0 7.155,3.203 7.155,7.155v407.247c0,3.952 -3.203,7.155 -7.155,7.155z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m266.859,466.778h-10.859v-421.557h10.859c3.952,0 7.155,3.203 7.155,7.155v407.247c0,3.952 -3.203,7.155 -7.155,7.155z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="M423.74,80.381m-60.946,0a60.946,60.946 0,1 1,121.892 0a60.946,60.946 0,1 1,-121.892 0" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m511.914,326.904 l-15.196,-128.529c-1.537,-13 -12.557,-22.794 -25.647,-22.794h-94.663c-13.09,0 -24.11,9.794 -25.647,22.794l-15.196,128.529c-0.857,7.248 4.805,13.618 12.104,13.618l21.612,0.081 -0.567,5.1 -3.536,31.799c-0.401,3.61 2.424,6.767 6.057,6.767h0.674l8.044,97.113c0.523,6.319 5.805,11.182 12.146,11.182h63.283c6.341,0 11.623,-4.862 12.146,-11.182l8.044,-97.113h0.674c3.632,0 6.458,-3.157 6.057,-6.767l-3.536,-31.799 -0.573,-5.181h21.618c7.297,0.001 12.959,-6.37 12.102,-13.618z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m467.197,241.673c-0.686,-6.172 -5.903,-10.841 -12.113,-10.841h-31.344,-31.344c-6.21,0 -11.427,4.669 -12.113,10.841l-15.105,135.829c-0.401,3.61 2.424,6.767 6.057,6.767h52.506,52.506c3.632,0 6.458,-3.157 6.057,-6.767z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m476.25,384.275h-18.605c3.628,0 6.45,-3.163 6.057,-6.77l-15.111,-135.836c-0.682,-6.171 -5.902,-10.832 -12.114,-10.832h18.605c6.212,0 11.432,4.662 12.114,10.832l15.111,135.836c0.393,3.607 -2.429,6.77 -6.057,6.77z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m484.684,80.376c0,33.664 -27.287,60.951 -60.941,60.951 -3.525,0 -6.977,-0.3 -10.336,-0.878 28.734,-4.9 50.605,-29.933 50.605,-60.073 0,-30.129 -21.871,-55.163 -50.605,-60.062 3.359,-0.579 6.811,-0.879 10.336,-0.879 33.654,0 60.941,27.287 60.941,60.941z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m475.568,384.275 l-8.041,97.107c-0.527,6.326 -5.809,11.184 -12.145,11.184h-20.672c6.336,0 11.618,-4.858 12.145,-11.184l8.041,-97.107z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m499.806,340.523h-20.61c7.277,-0.041 12.899,-6.388 12.042,-13.623l-15.194,-128.528c-1.53,-12.992 -12.558,-22.791 -25.644,-22.791h20.672c13.085,0 24.114,9.799 25.644,22.791l15.193,128.528c0.858,7.256 -4.796,13.623 -12.103,13.623z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="M89.783,80.381m-60.946,0a60.946,60.946 0,1 1,121.892 0a60.946,60.946 0,1 1,-121.892 0" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m153.745,175.581h-127.924c-14.26,0 -25.821,11.561 -25.821,25.821v126.932c0,6.731 5.457,12.188 12.188,12.188h22.141v3.667,136.187c0,6.731 5.457,12.188 12.188,12.188h17.063c6.731,0 12.188,-5.457 12.188,-12.188v-82.322c0,-7.613 6.171,-13.784 13.784,-13.784h0.463c7.613,0 13.784,6.172 13.784,13.784v82.322c0,6.731 5.457,12.188 12.188,12.188h17.063c6.731,0 12.188,-5.457 12.188,-12.188v-136.186,-3.667h22.141c6.731,0 12.188,-5.456 12.188,-12.188v-126.933c0,-14.26 -11.561,-25.821 -25.822,-25.821z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m150.728,80.376c0,33.664 -27.287,60.951 -60.941,60.951 -3.886,0 -7.69,-0.362 -11.37,-1.065 28.217,-5.323 49.571,-30.109 49.571,-59.887 0,-29.768 -21.354,-54.553 -49.571,-59.876 3.68,-0.703 7.483,-1.065 11.37,-1.065 33.654,0.001 60.941,27.288 60.941,60.942z" />
-    <path
-        android:fillColor="?attr/colorPrimary"
-        android:pathData="m179.565,201.4v126.936c0,6.729 -5.457,12.186 -12.186,12.186h-22.14s11.587,-5.457 11.587,-12.186v-126.936c0,-14.253 -11.556,-25.819 -25.819,-25.819h22.739c14.264,0 25.819,11.566 25.819,25.819z" />
-</vector>
